// frontend/src/lib/api/roborev/generated/models/daemonStatus.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { AutoDesignStatus } from "./autoDesignStatus.ts";

export interface DaemonStatus {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  active_workers: number;
  address?: string;
  applied_jobs: number;
  auto_design?: AutoDesignStatus;
  canceled_jobs: number;
  completed_jobs: number;
  /** @minimum 0 */
  config_reload_counter?: number;
  config_reloaded_at?: string;
  failed_jobs: number;
  machine_id?: string;
  max_workers: number;
  network?: string;
  port?: number;
  queue_paused: boolean;
  queued_jobs: number;
  rebased_jobs: number;
  running_jobs: number;
  skipped_jobs: number;
  version: string;
}